[Chore] ScriptCell 파일 분리 #95
